+++ flutter_geo_currencies/lib/src/implementations/live.dart
@@ -1,10 +1,8 @@
 import 'package:intl/intl.dart';
 
 import '../../geo_currencies.dart';
-import '../models/conversion_data.dart';
-import '../models/currency_data.dart';
 import '../services/exchange_rate.dart';
-import '../services/open_street_map_utils.dart';
+import '../services/open_street_map.dart';
 import 'common/functions.dart';
 
 /// Represents country code associated with his information.
@@ -995,8 +993,6 @@ const Map<String, Map<String, String>> _countryCurrencyData = {
 class GeoCurrenciesLiveImplementation implements GeoCurrencies {
   final GeoCurrenciesConfig _config;
 
-  final GeocodingOpenStreetMap _geocoding = GeocodingOpenStreetMap();
-
   /// Constructs a new [GeoCurrenciesLiveImplementation].
   GeoCurrenciesLiveImplementation(this._config);
 
@@ -1014,7 +1010,7 @@ class GeoCurrenciesLiveImplementation implements GeoCurrencies {
     required double latitude,
     required double longitude,
   }) async {
-    final countryCode = await _geocoding.getCountryCode(
+    final countryCode = await GeocodingOpenStreetMap.getCountryCode(
       latitude: latitude,
       longitude: longitude,
     );
@@ -1044,10 +1040,21 @@ class GeoCurrenciesLiveImplementation implements GeoCurrencies {
         amount: amount,
         currencyCodeIso4217: currencyCodeIso4217,
         config: _config,
+        formatWithSymbol: true,
       );
 
   @override
-  Future<ConversionData?> convertAmount({
+  String formatAmountWithCurrencyCode(
+          {required num amount, required String currencyCodeIso4217}) =>
+      formatAmount(
+        amount: amount,
+        currencyCodeIso4217: currencyCodeIso4217,
+        config: _config,
+        formatWithSymbol: false,
+      );
+
+  @override
+  Future<ConversionData?> convertAmountWithCurrenciesCodes({
     required num amount,
     required String fromCurrencyCodeIso4217,
     required String toCurrencyCodeIso4217,
@@ -1059,6 +1066,33 @@ class GeoCurrenciesLiveImplementation implements GeoCurrencies {
     return await ExchangeRate.convertAmount(
       amount: amount,
       config: _config,
+      fromCurrencyCodeIso4217: fromCurrencyCodeIso4217,
+      toCurrencyCodeIso4217: toCurrencyCodeIso4217,
+      currencyConversionData: currencyConversionData,
+    );
+  }
+
+  @override
+  AmountConvertedData convertAmountWithRate(
+          {required num amount,
+          required num rate,
+          required String toCurrencyCodeIso4217}) =>
+      convertAmount(
+          amount: amount,
+          config: _config,
+          rate: rate,
+          toCurrencyCodeIso4217: toCurrencyCodeIso4217);
+
+  @override
+  Future<RateData?> getRate({
+    required String fromCurrencyCodeIso4217,
+    required String toCurrencyCodeIso4217,
+  }) async {
+    final currencyConversionData = await ExchangeRate.getConversionData(
+      fromCurrencyCodeIso4217: fromCurrencyCodeIso4217,
+    );
+
+    return await ExchangeRate.getRate(
       fromCurrencyCodeIso4217: fromCurrencyCodeIso4217,
       toCurrencyCodeIso4217: toCurrencyCodeIso4217,
       currencyConversionData: currencyConversionData,