+++ flutter_geo_currencies/lib/src/models/amount_converted_data.dart
@@ -1,10 +1,5 @@
-import 'currency_conversion_data.dart';
-
-/// Representation of the conversion data.
-class ConversionData {
-  /// The stored key ref for the [error] property.
-  static const keyError = 'error';
-
+/// Representation of the amount converted data.
+class AmountConvertedData {
   /// The stored key ref for the [baseAmount] property.
   static const keyBaseAmount = 'baseAmount';
 
@@ -19,29 +14,14 @@ class ConversionData {
   static const keyFormattedAmountConvertedWithCurrencySymbol =
       'formattedAmountConvertedWithCurrencySymbol';
 
-  /// The stored key ref for the [fromCurrencyCodeIso4217] property.
-  static const keyFromCurrencyCodeIso4217 = 'fromCurrencyCodeIso4217';
-
   /// The stored key ref for the [toCurrencyCodeIso4217] property.
   static const keyToCurrencyCodeIso4217 = 'toCurrencyCodeIso4217';
-
-  /// The stored key ref for the [succeeded] property.
-  static const keySucceeded = 'succeeded';
-
-  /// The stored key ref for the [currencyConversionData] property.
-  static const keyCurrencyConversionData = 'currencyConversionData';
 
   /// The stored key ref for the [rate] property.
   static const keyRate = 'rate';
 
-  /// The error.
-  final String? error;
-
   /// The base amount.
   final num? baseAmount;
-
-  /// Wether the conversion is succeed.
-  final bool succeeded;
 
   /// The amount converted.
   final num? amountConverted;
@@ -55,44 +35,30 @@ class ConversionData {
   /// The currency code iso 4217 to which we must convert.
   final String? toCurrencyCodeIso4217;
 
-  /// The currency code iso 4217 from which to convert.
-  final String? fromCurrencyCodeIso4217;
-
-  /// The currency conversion data.
-  final CurrencyConversionData? currencyConversionData;
-
   /// The rate.
   final num? rate;
 
-  /// Constructs a  new [ConversionData] from [Map] object.
-  ConversionData.fromJson(
+  /// Constructs a  new [AmountConvertedData] from [Map] object.
+  AmountConvertedData.fromJson(
     Map<String, dynamic> json,
   )   : baseAmount = json[keyBaseAmount],
         amountConverted = json[keyAmountConverted],
-        fromCurrencyCodeIso4217 = json[keyFromCurrencyCodeIso4217],
         toCurrencyCodeIso4217 = json[keyToCurrencyCodeIso4217],
-        succeeded = json[keySucceeded],
         formattedAmountConvertedWithCurrencySymbol =
             json[keyFormattedAmountConvertedWithCurrencySymbol],
         formattedAmountConvertedWithCurrencyCode =
             json[keyFormattedAmountConvertedWithCurrencyCode],
-        currencyConversionData = json[keyCurrencyConversionData],
-        rate = json[keyRate],
-        error = json[keyError];
+        rate = json[keyRate];
 
   /// Converts class to json.
   Map<String, dynamic> toJson() => {
-        keyError: error,
         keyBaseAmount: baseAmount,
         keyAmountConverted: amountConverted,
-        keyFromCurrencyCodeIso4217: fromCurrencyCodeIso4217,
         keyToCurrencyCodeIso4217: toCurrencyCodeIso4217,
         keyFormattedAmountConvertedWithCurrencyCode:
             formattedAmountConvertedWithCurrencyCode,
         keyFormattedAmountConvertedWithCurrencySymbol:
             formattedAmountConvertedWithCurrencySymbol,
-        keyCurrencyConversionData: currencyConversionData,
         keyRate: rate,
-        keySucceeded: succeeded,
       };
 }