+++ flutter_geo_currencies/CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+
+## Version 0.0.1:
+- Added `getCurrencyDataByCoordinate` function to retrieve currency data for a specific location specified by latitude and longitude.
+
+## Version 0.0.2:
+- Added `formatAmountWithCurrencySymbol` function to format the given amount with the currency symbol of the specified currency code (ISO 4217).
+- Added `formatAmountWithCurrencyCode` function to format the given amount with the currency code (ISO 4217) into a readable string.
+
+## Version 0.0.3:
+- Added `convertAmountWithCurrenciesCodes` function to convert the given amount from one currency code (ISO 4217) to another using currency codes.
+
+## Version 0.0.4:
+- Added `getRate` function to retrieve the conversion rate between two currency codes (ISO 4217).
+- Added `convertAmountWithRate` function to convert the given amount using the provided rate to the target currency code (ISO 4217).